Training works as intended
Upon training, each stat is increased by a random number between 1 to 5. The token being trained is removed, and a new token with the new stats is added. The UI updates accordingly.

Problem:
Same as that of marketplace, training a token requires successfully sending a signed transaction. Therefore, training currently does not work with any other accounts than the head authority unless the private key is manually changed in `config.js`.